refactor row-serde and update comments
